move code to StringUtils::CommaSeparatedStringIterator
