+++ src/hotspot/share/utilities/stringUtils.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -49,6 +49,49 @@ public:
   // eg. str "_abc____def__" would match pattern "abc*def".
   // The matching is case insensitive.
   static bool is_star_match(const char* star_pattern, const char* str);
+
+  class CommaSeparatedStringIterator {
+  private:
+    char* _token;
+    char* _saved_ptr;
+    char* _list;
+
+  public:
+    CommaSeparatedStringIterator(ccstrlist option) {
+      // Immediately make a private copy of option, and
+      // replace spaces and newlines with comma.
+      _list = (char*) canonicalize(option);
+      _saved_ptr = _list;
+      _token = strtok_r(_saved_ptr, ",", &_saved_ptr);
+    }
+
+    ~CommaSeparatedStringIterator() {
+      FREE_C_HEAP_ARRAY(char, _list);
+    }
+
+    const char* operator*() const { return _token; }
+
+    CommaSeparatedStringIterator& operator++() {
+      _token = strtok_r(nullptr, ",", &_saved_ptr);
+      return *this;
+    }
+
+    ccstrlist canonicalize(ccstrlist option_value) {
+      char* canonicalized_list = NEW_C_HEAP_ARRAY(char, strlen(option_value) + 1, mtCompiler);
+      int i = 0;
+      char current;
+      while ((current = option_value[i]) != '\0') {
+        if (current == '\n' || current == ' ') {
+          canonicalized_list[i] = ',';
+        } else {
+          canonicalized_list[i] = current;
+        }
+        i++;
+      }
+      canonicalized_list[i] = '\0';
+      return canonicalized_list;
+    }
+  };
 };
 
 #endif // SHARE_UTILITIES_STRINGUTILS_HPP
